Moved foreign_object models.py into a module.
