+++ tests/foreign_object/models/__init__.py
@@ -1,0 +1,9 @@
+from .article import (
+    Article, ArticleIdea, ArticleTag, ArticleTranslation, NewsArticle,
+)
+from .person import Country, Friendship, Group, Membership, Person
+
+__all__ = [
+    'Article', 'ArticleIdea', 'ArticleTag', 'ArticleTranslation', 'Country',
+    'Friendship', 'Group', 'Membership', 'NewsArticle', 'Person',
+]
